Move rpath appending logic into `repair`
